Feature/sprint2

+++ database/migrations/2023_02_14_150559_create_beds_table.php
@@ -21,7 +21,7 @@ class CreateBedsTable extends Migration
             $table->unsignedBigInteger('department_id');
             $table->float('charge');
             $table->string('notes')->nullable();
-            $table->float('status')->default(0);
+            $table->tinyInteger('status')->default(0);
             $table->timestamps();
             $table->softDeletes();
